safari-600.7-branch/LayoutTests/html5lib/generated/run-webkit01-write-expected.txt

CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
../resources/webkit01.dat: PASS
